+++ pkgs/ffigen/test/config_tests/json_schema_test.dart
@@ -7,11 +7,8 @@ import 'dart:io';
 
 import 'package:ffigen/ffigen.dart';
 import 'package:ffigen/src/strings.dart' as strings;
-import 'package:file/local.dart';
-import 'package:glob/glob.dart';
 import 'package:json_schema/json_schema.dart';
 import 'package:test/test.dart';
-import 'package:yaml/yaml.dart';
 
 import '../test_utils.dart';
 
@@ -45,44 +42,6 @@ void main() {
     test('Valid json schema', () {
       expect(jsonSchema, isNot(null));
     });
-
-    // Find all ffigen config files in the repo.
-    final configYamlGlob = Glob('**config.yaml');
-    final configYamlFiles = configYamlGlob.listFileSystemSync(
-      const LocalFileSystem(),
-      root: packagePathForTests,
-    );
-    test('$configYamlGlob files not empty', () {
-      expect(configYamlFiles.isNotEmpty, true);
-    });
-
-    final sharedBindingsConfigYamlGlob = Glob(
-      'example/shared_bindings/ffigen_configs/**.yaml',
-    );
-    final sharedBindingsConfigYamlFiles = sharedBindingsConfigYamlGlob
-        .listFileSystemSync(const LocalFileSystem(), root: packagePathForTests);
-    test('$sharedBindingsConfigYamlGlob files not emty', () {
-      expect(sharedBindingsConfigYamlFiles.isNotEmpty, true);
-    });
-
-    final allConfigFiles = configYamlFiles + sharedBindingsConfigYamlFiles;
-
-    for (final fe in allConfigFiles) {
-      test('validate config file: ${fe.path}', () {
-        final yamlDoc = loadYaml(File(fe.absolute.path).readAsStringSync());
-        final validationResult = jsonSchema.validate(yamlDoc);
-        expect(
-          validationResult.errors.isEmpty,
-          true,
-          reason: 'Schema Errors: ${validationResult.errors}',
-        );
-        expect(
-          validationResult.warnings.isEmpty,
-          true,
-          reason: 'Schema Warnings: ${validationResult.errors}',
-        );
-      });
-    }
 
     test('Bare minimal input', () {
       expect(