+++ pkgs/ffigen/test/example_tests/cjson_example_test.dart
@@ -6,19 +6,22 @@ import 'package:ffigen/src/header_parser.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 
+import '../../example/c_json/tool/ffigen.dart' as cjson_example;
 import '../test_utils.dart';
 
 void main() {
   group('cjson_example_test', () {
     test('c_json', () async {
-      final config = testConfigFromPath(
-        path.join(packagePathForTests, 'example', 'c_json', 'config.yaml'),
+      final packageRoot = path.join(packagePathForTests, 'example', 'c_json/');
+      final context = testContext(
+        cjson_example.getConfig(Uri.file(packageRoot)),
       );
-      final context = testContext(config);
       final library = parse(context);
 
       await matchLibraryWithExpected(context, library, 'example_c_json.dart', [
-        config.output.dart.path.toFilePath(),
+        'example',
+        'c_json',
+        'cjson_generated_bindings.dart',
       ]);
     });
   });